+++ src/requests/redpepper/requests/__init__.py
@@ -1,0 +1,5 @@
+from redpepper.common.requests import RequestError
+
+__all__ = [
+    "RequestError",
+]